backend/requirements/extra.txt
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
channels[daphne]
jwt